pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "analog-ptq"
version = "0.1.0"
description = "A modular post-training quantization pipeline for LLMs with customizable algorithms and integrated benchmarking"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.8"
authors = [
    {name = "Roy Dayan"}
]
keywords = ["quantization", "llm", "gptq", "post-training-quantization", "transformers"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0,<2.5.0",
    "transformers>=4.35.0,<4.47.0",
    "accelerate>=0.24.0,<1.1.0",
    "datasets>=2.14.0,<3.2.0",
    "lm-eval>=0.4.0,<0.5.0",
    "pydantic>=2.0.0,<3.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0,<14.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
analog-ptq = "analog_ptq.pipeline.runner:main"

[project.urls]
Homepage = "https://github.com/roydayan/analog-ptq"
Repository = "https://github.com/roydayan/analog-ptq"

[tool.setuptools.packages.find]
where = ["."]
include = ["analog_ptq*"]

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]

[tool.ruff]
line-length = 100
target-version = "py38"
select = ["E", "F", "W", "I", "N", "UP", "B", "C4"]
ignore = ["E501"]
